+++ src/main/java/ra/services/ticket/ITicketService.java
@@ -5,5 +5,5 @@ import ra.dtos.TicketDTO;
 import java.util.List;
 
 public interface ITicketService {
-    List<TicketDTO> getTicketsByUserId(Integer userId);
+    List<TicketDTO> getTicketsByUserId(Long userId);
 }
